# docs/ROLES_API_CONTRACTS.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CASA roles governance contracts",
    "version": "1"
  },
  "paths": {
...
    "/api/admin/courses": {
      "get": {
        "tags": [
          "admin-content"
        ],
        "summary": "Admin List Courses",
        "operationId": "admin_list_courses_api_admin_courses_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "School Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ContentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCourseListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Admin Create Course",
        "operationId": "admin_create_course_api_admin_courses_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminCourseIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCourseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/courses/preview-pdf": {
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Admin Preview Pdf",
        "description": "Analyse un PDF et renvoie ce que l'import produirait, sans rien créer.\n\nPas de `school_id` ni de session : la prévisualisation ne touche à rien\nen base, et n'a donc besoin ni de l'une ni de l'autre.",
        "operationId": "admin_preview_pdf_api_admin_courses_preview_pdf_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_admin_preview_pdf_api_admin_courses_preview_pdf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/courses/import-pdf": {
      "post": {
        "tags": [
          "admin-content"
        ],
        "summary": "Admin Import Pdf",
        "description": "Importe un PDF.\n\n`create_course=false` verse le document au corpus documentaire sans créer\nde cours : la réponse ne porte alors ni `course_id` ni `lesson_id`.",
        "operationId": "admin_import_pdf_api_admin_courses_import_pdf_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_admin_import_pdf_api_admin_courses_import_pdf_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminCourseIn": {
        "properties": {
          "pathway_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pathway Id"
          },
          "school_id": {
            "type": "string",
            "title": "School Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus",
            "default": "DRAFT"
          }
        },
        "type": "object",
        "required": [
          "school_id",
          "title"
        ],
        "title": "AdminCourseIn"
      },
      "AdminCourseListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminCourseOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AdminCourseListResponse"
      },
      "AdminCourseOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "school_id": {
            "type": "string",
            "title": "School Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "duration_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Min"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "school_id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AdminCourseOut"
      },
...
      "Body_admin_import_pdf_api_admin_courses_import_pdf_post": {
        "properties": {
          "school_id": {
            "type": "string",
            "title": "School Id"
          },
          "pathway_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pathway Id"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "create_course": {
            "type": "boolean",
            "title": "Create Course",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "school_id",
          "file"
        ],
        "title": "Body_admin_import_pdf_api_admin_courses_import_pdf_post"
      },
      "Body_admin_preview_pdf_api_admin_courses_preview_pdf_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "school_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Id"
          },
          "pathway_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pathway Id"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_admin_preview_pdf_api_admin_courses_preview_pdf_post"
      },
...
      "ContentStatus": {
        "type": "string",
        "enum": [
          "DRAFT",
          "PUBLISHED",
          "ARCHIVED"
        ],
        "title": "ContentStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PdfImportResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "course_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Id"
          },
          "lesson_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lesson Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "pages_extracted": {
            "type": "integer",
            "title": "Pages Extracted"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          },
          "report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "title",
          "pages_extracted"
        ],
        "title": "PdfImportResponse"
      },
      "PdfPreviewBlockOut": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "preview": {
            "type": "string",
            "title": "Preview"
          },
          "items": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "confidence",
          "preview"
        ],
        "title": "PdfPreviewBlockOut"
      },
      "PdfPreviewResponse": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          },
          "report": {
            "additionalProperties": true,
            "type": "object",
            "title": "Report"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/PdfPreviewSectionOut"
            },
            "type": "array",
            "title": "Sections",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "pages",
          "report"
        ],
        "title": "PdfPreviewResponse"
      },
      "PdfPreviewSectionOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "page_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "blocks": {
            "items": {
              "$ref": "#/components/schemas/PdfPreviewBlockOut"
            },
            "type": "array",
            "title": "Blocks",
            "default": []
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/PdfPreviewSectionOut"
            },
            "type": "array",
            "title": "Children",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "level",
          "confidence"
        ],
        "title": "PdfPreviewSectionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}